Prepare for reorganization merge
